Fixed #1: StratusAssets now are equatable
While the issue said "Comparible", it was talking about being able to state that two assets were equal readily.  Thus what was meant was IEquatable.

I chose to use Linq's SequenceEqual in the data for simplicity.  Yes I know there's faster options, but the target for this is unit tests.  If you can prove a need for speed, put in a patch and I'll consider it.

The time value is using equality because this is about equality: they should be identical to the last tick.