Heavy simplification of direct configuration API
Insteead of passing an array of array of server config instances, now you just pass an array of array of server instances.  Internally ChattelCache will duplicate the lists so that outside software can't manipulate the list after the fact.  As long as the server instance classes are built to have immutable properties this should be fairly safe.  Of course if someone disposes of the server instance before Chattel is done with it, then they will cause an NRE.