Doesnt compile or work. Attempting to make the Nini ctor general
The idea was to have a way that new types could be added similarly to the direct ctor.  However there's an ownership problem with the IDisposable server instances: does the calling code own them and have to dispose of the instances or is the ChattelConfiguration responsible?  In the past I'd had the ChattelConfiguration responsible, but it failed to fully manange that responsiblity.  Since I only ever used it once in any given program, and it lasted the full lifetime of the program, I never ran across the leakage that results in.  Until the unit test work, and the need to mock some of this.  Now I see how bad the design was.

But I may want to revisit this again, hence shoving this mess into a branch.