BREAKING CHANGE: Removed the ChattelConfiguration Nini ctor
The Nini config ctor causes not only an extra dependency, but causes an object ownership problem that results in both a potially heavy memory leak and a problem with the server connectors not closing properly because they never got disposed.

It's now entirely up to the application to handle the configuration.  Which means a pile of work for me, but maybe I'll figure out a clean way to handle it.  I've shoved start of trying to cross the fine line code off into a new branch to moulder.  Not sure it's worthwhile, but I also don't want to lose the idea just yet.